requirements.txt
streamlit>=1.32.0
google-generativeai>=0.5.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pymupdf>=1.24.0
numpy>=1.24.0
